use text files to load wave and paths
a wave is a list of enemies with defied starting positions and path
type. Path is a list of points.

user can define a wave and the path the enemies will take using a text
file in internal storage. seems to work fine, it's important to use the
correct encoding so far i'm sticking with western europe windows 1252